+++ examples/skip.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    05 Feb 2025, 15:43:34
 //  Last edited:
-//    06 Feb 2025, 15:32:45
+//    13 Feb 2025, 15:30:38
 //  Auto updated?
 //    Yes
 //
@@ -14,9 +14,10 @@
 
 #![allow(unused)]
 
+use std::cmp::Ordering;
 use std::hash::{DefaultHasher, Hasher as _};
 
-use better_derive::{Debug, Hash, PartialEq};
+use better_derive::{Debug, Hash, PartialEq, PartialOrd};
 
 
 #[inline]
@@ -29,7 +30,7 @@ fn hash<T: std::hash::Hash>(obj: T) -> u64 {
 
 
 // If you use `cargo expand --example skip`, you can see the bounds on `u32` aren't even derived!
-#[derive(Debug, Hash, PartialEq)]
+#[derive(Debug, Hash, PartialEq, PartialOrd)]
 pub struct Foo {
     // This is interesting!
     bar: String,
@@ -37,6 +38,7 @@ pub struct Foo {
     #[debug(skip)]
     #[hash(skip)]
     #[partial_eq(skip)]
+    #[partial_ord(skip)]
     baz: u32,
     // Note: you can also do this!
     #[better_derive(skip)]
@@ -50,5 +52,6 @@ fn main() {
     let foo2 = Foo { bar: "Hello, world!".into(), baz: 43, quz: 421 };
     assert_eq!(format!("{:?}", foo1), "Foo { bar: \"Hello, world!\" }");
     assert!(foo1 == foo2);
+    assert_eq!(foo1.partial_cmp(&foo2), Some(Ordering::Equal));
     assert!(hash(foo1) == hash(foo2));
 }